distsqlrun: ensure errors are wrapped, not rewritten
This an oversight in #35821 - all the errors in `inbound.go` use
`errors.Wrap`, this one must too.

(This and the rest of the file should really use `pgerror.Wrap`, which
in combination with #36023 would be innocuous. But that's an aside.)
In either case, all the functionality in this file requires wrapping,
not flattening the error into a string.

Release note: None